requirements.txt
folium==0.14.0
matplotlib==3.7.3
numpy==1.24.3
requests==2.31.0
shapely==2.0.1
geopy==2.3.0
openrouteservice==2.3.3
python-dotenv==1.0.0
tqdm==4.66.1 